Revert logic of overlapping stats calculation for not target flows.

+++ src-java/flowhs-topology/flowhs-storm-topology/src/test/java/org/openkilda/wfm/topology/flowhs/validation/FlowValidatorTest.java
@@ -271,13 +271,13 @@ public class FlowValidatorTest {
 
     @Test(expected = InvalidFlowException.class)
     public void failIfMaxLatencyTier2HigherThanMaxLatencyTest() throws InvalidFlowException {
-        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2((long) 1000, (long) 500);
+        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2(1000L, 500L);
         flowValidator.checkMaxLatencyTier(flow);
     }
 
     @Test(expected = InvalidFlowException.class)
     public void failIfMaxLatencyTier2butMaxLatencyIsNullTest() throws InvalidFlowException {
-        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2(null, (long) 500);
+        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2(null, 500L);
         flowValidator.checkMaxLatencyTier(flow);
     }
 
@@ -323,13 +323,13 @@ public class FlowValidatorTest {
 
     @Test(expected = InvalidFlowException.class)
     public void failIfMaxLatencyTier2HigherThanMaxLatency() throws InvalidFlowException {
-        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2((long) 1000, (long) 500);
+        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2(1000L, 500L);
         flowValidator.checkMaxLatencyTier(flow);
     }
 
     @Test(expected = InvalidFlowException.class)
     public void failIfMaxLatencyTier2butMaxLatencyIsNull() throws InvalidFlowException {
-        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2(null, (long) 500);
+        RequestedFlow flow = getTestRequestWithMaxLatencyAndMaxLatencyTier2(null, 500L);
         flowValidator.checkMaxLatencyTier(flow);
     }
